Changed the layout of hooh Windows
