--- Assets/_ModdingToolScripts/Editor/ModdingComponentInitializer.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: fc31227ec5544bc7bb526edd81cab31e
-timeCreated: 1593302211
