+++ Assets/_ModdingToolScripts/asm_DebugComponents/Scripts/Preview/AIHSPresetLoader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3e8f04fd8ef528047bddce505eba7538
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
