+++ Assets/_ModdingToolScripts/asm_Loader/MapHelper.asmdef
@@ -1,7 +1,8 @@
 {
     "name": "NewAssembly",
     "references": [
-        "BundleManager"
+        "BundleManager",
+        "BakeryRuntimeAssembly"
     ],
     "optionalUnityReferences": [],
     "includePlatforms": [],
