+++ Assets/_ModdingToolScripts/asm_Packer/Editor/HoohToolWindow/XMLGenerator.cs
@@ -1,0 +1,13 @@
+﻿using hooh_ModdingTool.asm_Packer.Editor;
+using UnityEditor;
+
+public partial class HoohTools
+{
+    public void DrawXMLGenerator(SerializedObject serializedObject)
+    {
+        WindowUtility.VerticalLayout(() =>
+        {
+            WindowUtility.Button("Create Mod Template", () => { });
+        });
+    }
+}
