+++ Assets/_ModdingToolScripts/asm_Packer/ModPackerModule/Structure/SideloaderMod/SideloaderMod.cs
@@ -50,7 +50,7 @@ namespace ModPackerModule.Structure.SideloaderMod
         private const string BuildTargetName = "build";
         private readonly Dictionary<string, int> _autoPathIndex = new Dictionary<string, int>();
         private readonly List<BundleBase> _bundleTargets;
-        private readonly Dictionary<string, GameInfo> _gameItems;
+        public readonly Dictionary<string, GameInfo> GameItems;
         private readonly XDocument _inputDocumentObject;
         private readonly List<IManifestData> _manifestData;
         public readonly string AssetDirectory;
@@ -98,7 +98,7 @@ namespace ModPackerModule.Structure.SideloaderMod
             };
 
             _bundleTargets = new List<BundleBase>();
-            _gameItems = new Dictionary<string, GameInfo>();
+            GameItems = new Dictionary<string, GameInfo>();
             Assets = new AssetInfo();
 
             ParseDocument(_inputDocumentObject.Root);
@@ -221,7 +221,7 @@ namespace ModPackerModule.Structure.SideloaderMod
         {
             var target = element.Attr("target", AssetInfo.BundleTargetDefault);
             Assets.RememberTarget(target);
-            var dictItemList = _gameItems.ContainsKey(target) ? _gameItems[target] : new GameInfo(target);
+            var dictItemList = GameItems.ContainsKey(target) ? GameItems[target] : new GameInfo(target);
 
             foreach (var item in element.Elements("item"))
                 switch (type)
@@ -337,7 +337,7 @@ namespace ModPackerModule.Structure.SideloaderMod
                         throw new Exception($"\"{type}\" is a not valid category.");
                 }
 
-            _gameItems[target] = dictItemList;
+            GameItems[target] = dictItemList;
         }
 
         private void ParseListData(in XElement document)
