+++ Assets/_ModdingToolScripts/asm_Packer/Utility/HoohWindowStyles.cs
@@ -1,29 +1,29 @@
 ﻿#if UNITY_EDITOR
+using System.Diagnostics;
 using UnityEditor;
 using UnityEngine;
 
 namespace Common
 {
-    public class Styles
+    public class HoohWindowStyles
     {
+        public static HoohWindowStyles Instance => _instance ?? (_instance = new HoohWindowStyles());
+        private static HoohWindowStyles _instance;
         public GUIStyle Button;
         public GUIStyle ButtonDark;
         public GUIStyle ButtonGood;
         public GUIStyle ButtonWarning;
         public GUIStyle Foldout;
-
         public GUIStyle Header;
-        private bool init;
         public GUIStyle Medium;
         public GUIStyle Title;
+        public GUIStyle ButtonTab;
 
         public static Color red => new Color(0.91f, 0.3f, 0.24f);
-
         public static Color green => new Color(0.18f, 0.8f, 0.44f);
 
-        public void Init()
+        private HoohWindowStyles()
         {
-            if (init) return;
             Medium = new GUIStyle(GUI.skin.label)
             {
                 alignment = TextAnchor.MiddleLeft,
@@ -41,13 +41,15 @@ namespace Common
             {
                 fontSize = 11, margin = new RectOffset(2, 2, 2, 2), padding = new RectOffset(5, 5, 5, 5), fixedHeight = 23
             };
-            ButtonDark = new GUIStyle(Button);
-            ButtonDark.normal.textColor = Color.white;
-            ButtonDark.active.textColor = Color.white;
-            ButtonDark.hover.textColor = Color.white;
+            ButtonDark = new GUIStyle(Button) {normal = {textColor = Color.white}, active = {textColor = Color.white}, hover = {textColor = Color.white}};
             ButtonWarning = new GUIStyle(Button);
             ButtonGood = new GUIStyle(Button);
-            init = true;
+            
+            ButtonTab = new GUIStyle(GUI.skin.button)
+            {
+                fixedHeight = 32,
+                margin = new RectOffset(0, 0, 0, 32)
+            };
         }
     }
 }
