+++ Assets/_ModdingToolScripts/asm_Utility.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 46b3c9e18ffa63d43a2f91e54f5961ae
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
