+++ Assets/_ModdingToolScripts/asm_Utility/FixBadAssets.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e9eb6e59272b00c49948fe0fdf595e78
+guid: c49f709fda589cb4f904d02374b860be
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
