+++ Assets/_ModdingToolScripts/asm_Utility/UnityMacros.asmdef
@@ -1,0 +1,3 @@
+﻿{
+	"name": "UnityMacros"
+}
