+++ Assets/_ModdingToolScripts/asm_Utility/UnityMacros.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 38e1364e1ed606f42a761a1f7616f538
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
